Clean up version formatting
